+++ tests/test_scavenging.py
@@ -1,8 +1,4 @@
-import sys
-
-import pytest
-
-from . import check, skip_if_not_has_async, v
+from . import check, v
 
 assert v  # Silence pyflakes.
 
@@ -73,7 +69,6 @@ b = foo(5)
     check(v.defined_funcs, ["foo"])
 
 
-@skip_if_not_has_async
 def test_async_function(v):
     v.scan(
         """\
@@ -85,7 +80,6 @@ async def foo():
     check(v.unused_funcs, ["foo"])
 
 
-@skip_if_not_has_async
 def test_async_method(v):
     v.scan(
         """\
@@ -451,7 +445,6 @@ class OtherClass:
     check(v.unused_funcs, ["other_func"])
 
 
-@skip_if_not_has_async
 def test_async_function_name_in_test_file(v):
     v.scan(
         """\
@@ -467,7 +460,6 @@ async def other_func():
     check(v.unused_funcs, ["other_func"])
 
 
-@skip_if_not_has_async
 def test_async_function_name_in_normal_file(v):
     v.scan(
         """\
@@ -620,7 +612,6 @@ a = 2
     check(v.unused_vars, ["a", "a"])
 
 
-@pytest.mark.skipif(sys.version_info < (3, 5), reason="requires Python 3.5+")
 def test_arg_type_annotation(v):
     v.scan(
         """\
@@ -636,7 +627,6 @@ def f(n: int) -> Iterable[int]:
     check(v.unused_imports, [])
 
 
-@pytest.mark.skipif(sys.version_info < (3, 6), reason="requires Python 3.6+")
 def test_var_type_annotation(v):
     v.scan(
         """\